Ikke sååå granulere filer
