+++ src/components/melding/NyMelding.tsx
@@ -5,7 +5,7 @@ import { Link } from '@tanstack/react-router';
 import { useAtomValue } from 'jotai';
 import type { ReactElement } from 'react';
 import AvsluttDialogEtterSending from 'src/components/melding/AvsluttDialogEtterSending';
-import { Oppgaveliste, OppgavelisteRadioKnapper } from 'src/components/melding/OppgavelisteRadioKnapper';
+import OppgavelisteRadioKnapper, { Oppgaveliste } from 'src/components/melding/OppgavelisteRadioKnapper';
 import { ValgForMeldingstype } from 'src/components/melding/ValgForMeldingstype';
 import { MeldingsType, VelgMeldingsType, meldingsTyperTekst } from 'src/components/melding/VelgMeldingsType';
 import VelgOppgaveliste from 'src/components/melding/VelgOppgaveliste';
